+++ app/src/main/java/com/svdroid/avivgrouptest/data/db/PropertiesEntity.kt
@@ -6,12 +6,12 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class PropertiesEntity(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     @ColumnInfo(name = "bedrooms") val bedrooms: Int?,
     @ColumnInfo(name = "rooms") val rooms: Int?,
     @ColumnInfo(name = "offer_type") val offerType: Int?,
-    @ColumnInfo(name = "area") val area: Float?,
-    @ColumnInfo(name = "price") val price: Float?,
+    @ColumnInfo(name = "area") val area: Double?,
+    @ColumnInfo(name = "price") val price: Double?,
     @ColumnInfo(name = "url") val url: String?,
     @ColumnInfo(name = "city") val city: String?,
     @ColumnInfo(name = "professional") val professional: String?,
